fcd: restore support for FUNcube Dongle and Pro+
Support for the original FUNcube Dongle used to live in GNU Radio's 
gr-fcd module, which was removed in version 3.8. As a result, gr-osmosdr 
lost support for both FUNcube Dongle and FUNcube Dongle Pro+.

The gr-fcdproplus out-of-tree module subsequently added support for the 
original FUNcube Dongle, meaning that it now supports both types. As a 
result, FUNcube support can easily be restored in gr-osmosdr. The 
#ifdefs are no longer required, since everything is in a single module 
now.